Restructured the project files. The mod code should be first citizen in the project root. The SEModCore sits to the side. This also help in keeps namespaces correct. Fixed issue in Scan, where /Ignore was not properly split client/server.
Updated SEModCore 4.1
Implemented granular logging, which shows all when debugging.
Corrected abstract methods, and sealed methods that should not be overwritten.
Changed the constants to initial parameters which must be passed in, this also helped seperate mod code from framework.
Fixed the ConnectionResponse as it wasn't been received correctly, causing the Client to continue requesting it.